requirements.txt
async-timeout==4.0.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
spotipy==2.24.0
urllib3==2.2.3
Werkzeug==3.0.4
ytmusicapi==1.8.1
zipp==3.20.2
